Add a wks file for 8G sd cards.
wic normally calcualtes the size of a partition based on the amont of stuff
installed plus an overhead factor. This hard codes the the size for an 8G sd
card. This is good and bad, bad in that it takes a long time to write, good
in that you do not need to figure out how to resize the partition.

Signed-off-by: Philip Balister <philip@balister.org>