recipes-support: Add swig-native to all GNU Radio OOT modules
OE uses per-package sysroots, which means that swig-native needs to
explicitly added to each OOT module, or it won't find a native swig
executable to build the SWIG subcomponents.
Note that omitting swig-native doesn't lead to a build failure, unless
OOTs are modified to fail on missing SWIG. The default behaviour is to
simply disable Python.

Signed-off-by: Martin Braun <martin.braun@ettus.com>