gnuradio: Fixed PACKAGECONFIG for grc
Move runtime dependencies of grc to RDEPENDS_pn-gnuradio-grc. Previously they
were configured as 3rd argument of PACKAGECONFIG[grc] which is incorrect because
they this value defines the build time dependencies.

At the same time, cleanup dependencies as well:

- There are no additional build time dependencies for grc because the CMakeLists.txt
only checks for the Python libs in case of NOT CMAKE_CROSSCOMPILING
- Remove python3-netserver from runtime dependencies

Signed-off-by: Joerg Hofrichter <joerg.hofrichter@ni.com>