gnuradio-oot.bbclass: First attempt to collect OOT build steps.
Recipes have been added by copying another recipe and hacking until
it works. As a result recipes are hard to fix consistently and we
are carrying things that are no longer needed. This attempts to get
this under control and make it easier for people to add recipes for
GNU Radio OOT modules.

Signed-off-by: Philip Balister <philip@balister.org>